+++ BrainMatrixGenerator/BrainMatrixGenerator_v2_transposed.c
@@ -292,7 +292,6 @@ void generate_coo_brain_matrix_for_pagerank(IntCOOMatrix *M_COO, long ind_start_
             source_type = neuron_types[j];
             //récupération de la probabilité de connexion source -> destination avec le type de neurone donné
             proba_connection = (*brain).brainPart[ind_part_source].probaConnection[source_type*(*brain).nb_part + ind_part_dest];
-            //DEL if (ind_part_dest == 0) {printf("[Neurone %li,%li] Tentative de création d'une connexion de la partie %i à la partie %i (type %i) avec une proba %f\n",i,j,ind_part_source,ind_part_dest,source_type,proba_connection);}
             proba_no_connection = 1 - proba_connection;
             random = random_between_0_and_1();
             //décision aléatoire, en prenant en compte l'abscence de connexion sur la diagonale de façon brute
@@ -381,12 +380,12 @@ int main(int argc, char **argv)
     long i,j,k; //pour les boucles
     long n;
     long long size;
-    int nb_zeros,nb_non_zeros,nb_non_zeros_local,*list_nb_non_zeros_local;
+    long nb_zeros,nb_non_zeros,nb_non_zeros_local,*list_nb_non_zeros_local;
     long *nb_connections_local_tmp,*nb_connections_tmp;
     int *neuron_types, *local_types;
 
     //allocation mémoire pour les nombres de 0 dans chaque sous matrice de chaque processus
-    if (debug) {list_nb_non_zeros_local = (int *)malloc(p * sizeof(int));}
+    if (debug) {list_nb_non_zeros_local = (long *)malloc(p * sizeof(long));}
 
     //allocation mémoire et initialisation d'une liste de taille "nombre de processus" contenant les pourcentages de 0 que l'on souhaite pour chaque bloc
     int *zeros_percentages = (int *)malloc(p * sizeof(int)); for (i=0;i<p;i++) {zeros_percentages[i] = 75;}
@@ -553,8 +552,8 @@ int main(int argc, char **argv)
     }
 
     nb_non_zeros_local = A_COO.len_values;
-    MPI_Allreduce(&nb_non_zeros_local, &nb_non_zeros, 1, MPI_INT, MPI_SUM, MPI_COMM_WORLD); //somme MPI_SUM de tout les nb_non_zeros_local dans nb_non_zeros
-    if (debug) {MPI_Allgather(&nb_non_zeros_local, 1, MPI_INT, list_nb_non_zeros_local, 1,  MPI_INT, MPI_COMM_WORLD);} //réunion dans chaque processus de tout les nombres de zéros de chaque bloc
+    MPI_Allreduce(&nb_non_zeros_local, &nb_non_zeros, 1, MPI_LONG, MPI_SUM, MPI_COMM_WORLD); //somme MPI_SUM de tout les nb_non_zeros_local dans nb_non_zeros
+    if (debug) {MPI_Allgather(&nb_non_zeros_local, 1, MPI_LONG, list_nb_non_zeros_local, 1,  MPI_LONG, MPI_COMM_WORLD);} //réunion dans chaque processus de tout les nombres de zéros de chaque bloc
 
     if ((debug || debug_cerveau) && my_rank == 0)
     {
